feat: 3단계 구현 + enum 수정

+++ src/main/java/domain/PurchasePrice.java
@@ -30,7 +30,7 @@ public class PurchasePrice {
         return purchasePrice / 1000;
     }
 
-    public float calculateProfit(int price) {
+    public float calculateProfit(long price) {
         return (float) price / purchasePrice;
     }
 }
